extract application menu
