+++ src/lib/applicationMenu.js
@@ -1,0 +1,46 @@
+import {app, Menu,} from 'electron';
+
+export function initApplicationMenu() {
+    if (process.platform !== 'darwin') return;
+    const template = [
+        {
+            label: app.getName(),
+            submenu: [
+                {role: 'about'},
+                {type: 'separator'},
+                {role: 'hide'},
+                {role: 'hideothers'},
+                {role: 'unhide'},
+                {type: 'separator'},
+                {role: 'quit'}
+            ]
+        },
+        {
+            label: 'Edit',
+            submenu: [
+                {role: 'undo'},
+                {role: 'redo'},
+                {type: 'separator'},
+                {role: 'cut'},
+                {role: 'copy'},
+                {role: 'paste'},
+                {role: 'delete'},
+                {role: 'selectall'}
+            ]
+        },
+        {
+            role: 'window',
+            submenu: [
+
+                {role: 'close'},
+                {role: 'minimize'},
+                {role: 'zoom'},
+                {type: 'separator'},
+                {role: 'front'},
+            ]
+        }
+    ];
+
+    const menu = Menu.buildFromTemplate(template);
+    Menu.setApplicationMenu(menu);
+}
